Apply jQuery rules after deferred history loads, as well
There is no reliable way to only recieve events for .load() events under
a specific event; .ajaxComplete triggers are set on the whole document.
As such, we must be sure to only apply the quote-selection click handler
to each link once -- otherwise, it will run multiple times in succession
for a single click, appending an additional &UpdateContent= for each
time that an AJAX query has run on the page.